requirements.txt
aiodns==3.2.0
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email-normalize==2.0.0
email_validator==2.2.0
fastapi==0.115.8
h11==0.14.0
idna==3.10
motor==3.7.0
pyasn1==0.6.1
pycares==4.5.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11
python-jose==3.3.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
